% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii.R
\name{ascii}
\alias{ascii}
\title{Create an ascii graphics device}
\usage{
ascii(filename = NULL, width = NULL, height = NULL,
  font_aspect = 0.5, verbosity = 0)
}
\arguments{
\item{filename}{if given, write ascii to this file, otherwise write to console}

\item{width, height}{dimensions of text output (in characters). If NULL then
auto-detect console width and set approximate aspect ratio at 1/2}

\item{font_aspect}{Character spacing horizontally and vertically are almost
never the same. On many terminals character spacing horizontally is
half the spacing vertically. This argument is only used if width
is autodetected and/or height is not specified. Default: 0.5}

\item{verbosity}{level. default: 0.  valid: 0, 1, 2}
}
\description{
Graphics primitives will be rendered using ascii characters.
}
