% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ascii_}
\alias{ascii_}
\title{Create an ascii graphics device}
\usage{
ascii_(filename, width, height, verbosity = 0L)
}
\arguments{
\item{filename}{if given, write ascii to this file, otherwise write to console.
If file is a format string e.g. "output%03i.txt", then this will be used
to create a unique filename if multiple pages are output within the same
\code{ascii()} call.}

\item{width, height}{dimensions of text output (in characters)}

\item{verbosity}{level. default: 0}
}
\description{
Create an ascii graphics device
}
