% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Create a descriptive graphics device}
\usage{
descriptive(verbosity = 1L, width = 400L, height = 300L, ...)
}
\arguments{
\item{verbosity}{level. default: 1}

\item{width, height}{nominal size of drawing area in pixels. default: width = 400, height = 300}

\item{...}{all other arguments ignored}
}
\description{
Create a descriptive graphics device
}
