% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{descriptive_}
\alias{descriptive_}
\title{Create a descriptive graphics device}
\usage{
descriptive_(verbosity = 1L, width = 400L, height = 300L)
}
\arguments{
\item{verbosity}{level. default: 1}

\item{width, height}{nominal size of drawing area in pixels.}
}
\description{
Create a descriptive graphics device
}
