\name{devout-package}
\alias{devout-package}
\alias{devout}
\docType{package}
\title{
  Non-standard graphics devices for R
}
\description{
  A collection of non-standard graphics devices. Currently includes ascii output
    and textual output.
}
\details{
  A collection of non-standard graphics devices. Currently includes ascii output
    and textual output.
}
\author{
Maintainer: mikefc <mikefc@coolbutuseless.com>
}
\keyword{ package }
