% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device-description.R
\name{get_default_device_description}
\alias{get_default_device_description}
\title{Get the default device decription used in the rdevice}
\usage{
get_default_device_description(width = 100, height = 100, startps = 12)
}
\arguments{
\item{width, height}{dimensions of device}

\item{startps}{starting pointsize}
}
\value{
list of device description information
}
\description{
Get the default device decription used in the rdevice
}
