% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{keep_alive_env}
\alias{keep_alive_env}
\title{Use a package-wide cache to keep the 'matrix' backing store for 'ascii()'
alive when only the pointer itself is kept with C++ code}
\format{An object of class \code{environment} of length 0.}
\usage{
keep_alive_env
}
\description{
Use a package-wide cache to keep the 'matrix' backing store for 'ascii()'
alive when only the pointer itself is kept with C++ code
}
\keyword{datasets}
