% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_list}
\alias{modify_list}
\title{Simple non-recursive version of \code{modifyList} which works with lists
and environemnts}
\usage{
modify_list(old, new)
}
\arguments{
\item{old, new}{list-like objects}
}
\description{
Simple non-recursive version of \code{modifyList} which works with lists
and environemnts
}
