% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdevice.R
\name{rdevice}
\alias{rdevice}
\title{Create a rdevice graphics device}
\usage{
rdevice(rfunction, ...)
}
\arguments{
\item{rfunction}{character string containing name of callback function in R}

\item{...}{all other named, non-NULL objects are passed into the device}
}
\description{
Inspired by: http://www.omegahat.net/RGraphicsDevice/overview.html
}
