% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdevice_}
\alias{rdevice_}
\title{Create a rdevice graphics device}
\usage{
rdevice_(rdata)
}
\arguments{
\item{rdata}{a list of information used on the R side}
}
\description{
Create a rdevice graphics device
}
