% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device-description.R
\name{sanitize_device_description}
\alias{sanitize_device_description}
\title{Sanitize the Device Description values}
\usage{
sanitize_device_description(dd)
}
\arguments{
\item{dd}{device description list of values}
}
\description{
Sanitize the Device Description values
}
