% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-callback.R
\name{verbose}
\alias{verbose}
\title{Verbose device}
\usage{
verbose(skip = c("mode", "strWidthUTF8", "metricInfo", "clip"), ...)
}
\arguments{
\item{skip}{names of device calls to ignore}

\item{...}{other parameters passed to the rdevice}
}
\description{
Simply prints all calls to the graphics driver
}
\details{
Uses \code{devout::rdevice()}.
}
