% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug-callback.R
\name{verbose_callback}
\alias{verbose_callback}
\title{Verbose debugging callback for the rdevice}
\usage{
verbose_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device function call}

\item{args}{arguments to device function call}

\item{state}{list of rdata, dd and gc. Some or all of which may be NULL}
}
\description{
Verbose debugging callback for the rdevice
}
