% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii.R
\name{ascii}
\alias{ascii}
\title{Graphics device for ASCII output}
\usage{
ascii(filename = NULL, width = NULL, height = NULL,
  font_aspect = 0.45, ...)
}
\arguments{
\item{filename}{If given, write ascii to this file, otherwise write to console.}

\item{width, height}{dimensions of text output (in characters). Default: NULL
(auto-detect)}

\item{font_aspect}{Character spacing horizontally and vertically are almost
never the same. On many terminals character resolution vertically is
half the resolution horizontally.  Adjust this value if circles don't
look right. Default: 0.45}

\item{...}{other parameters passed to the rdevice}
}
\description{
Graphics primitives will be rendered using ASCII characters.
}
\details{
Uses \code{devout::rdevice()}.
}
