% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdevice.R
\name{rdevice}
\alias{rdevice}
\title{Create an rdevice graphics device}
\usage{
rdevice(rfunction, ..., device_name = "rdevice")
}
\arguments{
\item{rfunction}{a function (preferred) or
a character string (soft-deprecated) containing name of callback function
which will handle the device calls.}

\item{...}{all other named, non-NULL arguments are passed into the device
as `rdata`}

\item{device_name}{name to use for the device. default: "rdevice"}
}
\description{
Inspired by: http://www.omegahat.net/RGraphicsDevice/overview.html
}
