% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdevice.R
\name{rdevice}
\alias{rdevice}
\title{Create an rdevice graphics device}
\usage{
rdevice(rfunction, ...)
}
\arguments{
\item{rfunction}{a function (preferred) or
a character string (soft-deprecated) containing name of callback function
which will handle the device calls.}

\item{...}{all other named, non-NULL arguments are passed into the device
as `rdata`}
}
\description{
Inspired by: http://www.omegahat.net/RGraphicsDevice/overview.html
}
