% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdevice.R
\name{sanitize_return_types}
\alias{sanitize_return_types}
\title{Sanitize the types of any explicit 'return' values expected by 'rdevice'}
\usage{
sanitize_return_types(device_call, state)
}
\arguments{
\item{device_call}{device call name}

\item{state}{return list}
}
\description{
Sanitize the types of any explicit 'return' values expected by 'rdevice'
}
